requirements.txt
# Portfolio Bot - Dependencies
# Python 3.11+

# =============================================================================
# Core LangChain / LangGraph
# =============================================================================
langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.0
langgraph>=0.2.0

# =============================================================================
# Vector Store
# =============================================================================
chromadb>=0.5.0

# =============================================================================
# API Framework
# =============================================================================
fastapi>=0.115.0
uvicorn>=0.32.0
httpx>=0.27.0

# =============================================================================
# Data Validation
# =============================================================================
pydantic>=2.0.0
pydantic-settings>=2.0.0

# =============================================================================
# Utilities
# =============================================================================
python-dotenv>=1.0.0

# =============================================================================
# Development (optional)
# =============================================================================
# black>=24.0.0
# isort>=5.0.0
# pre-commit>=3.0.0

# =============================================================================
# Graph Visualization (optional)
# =============================================================================
# grandalf>=0.8  # For graph layout
